Update hello_vsi doc to be consistent with Corstone_SSE-310 gated example
I went through the examples and noticed a couple of minor inconsistencies. Namely, the `Build` and `Run` sections refer to a Corstone 310, gated build, while the commands are for a Corstone 300, event build. I tried both and they worked, but I had to modify both build and run commands slightly. This commit updates the doc and commands within to use flow type `gated` and build target `Corstone_SSE-310, assuming this is the intention.